+++ dtsai-20260401/alibabacloud_dtsai20260401/models/_create_doc_parser_job_response_body.py
@@ -14,11 +14,17 @@ class CreateDocParserJobResponseBody(DaraModel):
         request_id: str = None,
         success: bool = None,
     ):
+        # The error code.
         self.error_code = error_code
+        # The error message.
         self.error_message = error_message
+        # The HTTP status code.
         self.http_status_code = http_status_code
+        # The document parsing task ID.
         self.job_id = job_id
+        # The request ID.
         self.request_id = request_id
+        # Indicates whether the request was successful.
         self.success = success
 
     def validate(self):