+++ dtsai-20260401/ChangeLog.md
@@ -1,3 +1,8 @@
+2026-08-06 Version: 1.0.2
+- Update API WebSearch: add request parameters UrlScopeDomains.
+- Update API WebSearch: add request parameters UrlScopeMode.
+
+
 2026-06-16 Version: 1.0.0
 - Generated python 2026-04-01 for DtsAI.